+++ sinks/dashboard/dashboard.py
@@ -28,7 +28,7 @@ from items.plot_3D_orientation import Orientation3DDashItem
 from items.plot_3D_position import Position3DDashItem
 from items.can_message_table import CanMsgTableDashItem
 from items.image_dash_item import ImageDashItem
-from items.label_dash_item import LabelDashItem
+from items.text_dash_item import TextDashItem
 
 
 # Custom class derived from QGraphicsView to capture mouse
